+++ src/Symfony/Component/Messenger/Bridge/AmazonSqs/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Allow auto-setup to create the queue when the DSN names the account the client already authenticates as
  * Add `AmazonSqsFairQueueStamp` to set a `MessageGroupId` on standard queues, enabling SQS fair queues
  * Add the `ssl` DSN option, superseding `sslmode`
+ * Add `AmazonSqsReceivedStamp::getSystemAttributes()` exposing the SQS system attributes of received messages
 
 7.4
 ---